+++ newton/_src/sim/__init__.py
@@ -24,7 +24,7 @@ from .enums import (
     JointTargetMode,
     JointType,
 )
-from .modal import ModalBasis, ModalGeneratorBeam, ModalGeneratorPOD, ModalGeneratorSampled
+from .modal import ModalBasis, ModalGeneratorBeam, ModalGeneratorFEM, ModalGeneratorPOD, ModalGeneratorSampled
 from .model import Model
 from .state import State
 
@@ -38,6 +38,7 @@ __all__ = [
     "JointType",
     "ModalBasis",
     "ModalGeneratorBeam",
+    "ModalGeneratorFEM",
     "ModalGeneratorPOD",
     "ModalGeneratorSampled",
     "Model",